extracted GameInterface delegate class from GUI, this new class handle the pop-up dialogs
